Refactor components
Components are allocated in a pool handled by the instantiated World.
Each component has its own pool.
Also, components are associated to an Entity. Each entity has pointers to its components.
Components interact with other components. \
At the moment the components are injected at the definition of the dependent components.