POOL - Implement double linked list for retrieving allocated items
The presence of the macro 'POOL_DOUBLE_LINK' establishes which pool's items retrieval implementation will be used.
If it is defined -> the pool will have two lists, one for free elements and another for allocated items -> could be faster but requires more RAM
If it is not defined -> the pool will have only a list for the free elements. To iterate over the pool items, it is needed a bool variable indicating whether the element is allocated or not.