Add PEntity APIs to exploit new components
'update' allows to add a callback to the entity that is called at each frame. Here the entity should be managed. The update callback takes a PEntity as parameter, this has many advantages: the user doesn't need to save references to entities; the update callback could be one for many entities, the entities will be recognized based on some state variable.

'collision' enable global collisions for the entity. This means that the entity could collide with other entities which have collision enabled. Also a callback can be passed, it will be triggered at every collision of the entity. Also the collision callback takes a PEntity as parameter, featuring the same advantages of udpate callbacks.

'collided' does not enable global collision. It checks if the entity has collided with the entity passed as parameter. This is useful for implementing one-to-many collision detections: in this scenario the many entities will check for collision with the one entity in their update method. This approach will save performance.

'getState' and 'setState' are needed to access the state variables of entities. When they are accessed, a state component is created and connected to the entity.